backend/character_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
redis==5.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
aiosqlite==0.20.0
